Final Commit
In a hurry, implemented game logic, for hitting and standing but switching could not be supported in time.
Did not have the time to prepare test cases for all of the winning conditions.
Implemented the game flow itself so that it starts and ends seamlessly.
Fixed several bugs throughout the program.